feat: TransactionManager 생성
