+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,26 +1,17 @@
 package com.techcourse.service;
 
 import com.techcourse.domain.User;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceUtils;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.SQLException;
+import org.springframework.transaction.support.TransactionManager;
 
 public class TxUserService implements UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(TxUserService.class);
-
     private final UserService userService;
-    private final DataSource dataSource;
+    private final TransactionManager transactionManager;
 
-    public TxUserService(final UserService userService, final DataSource dataSource) {
+    public TxUserService(final UserService userService, final TransactionManager transactionManager) {
         this.userService = userService;
-        this.dataSource = dataSource;
+        this.transactionManager = transactionManager;
     }
 
     public User findById(final long id) {
@@ -32,26 +23,15 @@ public class TxUserService implements UserService {
     }
 
     public void changePassword(final long id, final String newPassword, final String createBy) {
-        final var connection = DataSourceUtils.getConnection(dataSource);
         try {
-            connection.setAutoCommit(false);
+            transactionManager.begin();
             userService.changePassword(id, newPassword, createBy);
-            connection.commit();
-        } catch (SQLException e) {
-            rollbackConnection(connection);
-            log.error("트랜잭션 수행 과정에서 오류가 발생했습니다 => Rollback");
+            transactionManager.commit();
+        } catch (DataAccessException e) {
+            transactionManager.rollback();
             throw new DataAccessException(e);
         } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-            TransactionSynchronizationManager.unbindResource(dataSource);
-        }
-    }
-
-    private void rollbackConnection(final Connection connection) {
-        try {
-            connection.rollback();
-        } catch (SQLException e) {
-            throw new DataAccessException(e);
+            transactionManager.end();
         }
     }
 }
