+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.transaction.support.TransactionManager;
 
 import javax.sql.DataSource;
 
@@ -36,7 +37,7 @@ class UserServiceTest {
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
         final var appUserService = new AppUserService(userDao, userHistoryDao);
-        final var userService = new TxUserService(appUserService, dataSource);
+        final var userService = new TxUserService(appUserService, new TransactionManager(dataSource));
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -52,7 +53,7 @@ class UserServiceTest {
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
         final var appUserService = new AppUserService(userDao, userHistoryDao);
-        final var userService = new TxUserService(appUserService, dataSource);
+        final var userService = new TxUserService(appUserService, new TransactionManager(dataSource));
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
